gemini/requirements.txt
google-genai
google-cloud-aiplatform